requirements.txt
Django==2.0.3
django-qr-code==0.4.1
ipython==6.2.1
Pillow==5.0.0
pip=9.0.3
psycopg2==2.7.4
psycopg2-binary==2.7.4